mac80211: iwlwifi: unlock 320MHz bandwidth
This patch mainly skip community patch.

Description of original one:
If a discrete NIC is connected to a PCIe link hat isn't at least
Gen3 (8.0 GT/s), then we cannot sustain 320 MHz traffic, so remove
that from EHT capabilities in that case.

While at it, also move setting 320 MHz beamformee to the right
place in the code so it's not set while not supporting 320 MHz.

In my noisy environtment I still have better results for 320 MHz:
  (6g/1/EHT320): ping pkt loss ap->sta: 0%, sta->ap: 0%
  (6g/1/EHT320): ap->sta: 1453 Mbits/sec, sta->ap: 2097 Mbits/sec

  (6g/1/EHT160): ping pkt loss ap->sta: 0%, sta->ap: 0%
  (6g/1/EHT160): ap->sta: 1266 Mbits/sec, sta->ap: 1706 Mbits/sec

Signed-off-by: Janusz Dziedzic <janusz.dziedzic@gmail.com>